+++ test/refresh_multi_tab.test.ts
@@ -3,16 +3,8 @@ import { login, waitToResolve } from "./utils";
 
 const testUrl = "http://localhost:5000/";
 
-test("refresh flow with multiple tabs", async ({
-  browserName,
-  context,
-  page,
-}) => {
-  // Skip webkit for advanced concurrency test
-  // test.skip(
-  //   browserName === "webkit",
-  //   "Cannot guarantee thread safety in webkit without broadcast channel"
-  // );
+test("refresh flow with multiple tabs", async ({ context, page }) => {
+  // This test is slow!
   test.slow();
 
   await page.goto(testUrl);
@@ -28,7 +20,7 @@ test("refresh flow with multiple tabs", async ({
 
   let previousToken = await page.textContent("span#access-token");
 
-  // Open many other pages
+  // Open many other tabs
   new Array(25).fill(null).forEach(async () => {
     const page = await context.newPage();
     await page.goto(testUrl);
@@ -37,11 +29,11 @@ test("refresh flow with multiple tabs", async ({
   // Test 10 successful rotations
   for (let i = 0; i < 10; i++) {
     // Wait for a rotation response from any tab
-    await context.waitForEvent(
-      "response",
-      (res) => res.url() === "http://localhost:5001/token"
-    );
-    await page.waitForTimeout(1000);
+    await context.waitForEvent("response", {
+      predicate: (res) => res.url() === "http://localhost:5001/token",
+      timeout: 30_000,
+    });
+    await page.waitForTimeout(1_000);
 
     // Check we have a new access token and are still logged in
     const nextToken = await page.textContent("span#access-token");